+++ qt/manageprofiles/__init__.py
@@ -65,97 +65,12 @@ import sshtools
 import logger
 import encfsmsgbox
 from manageprofiles import schedulewidget
+from manageprofiles.sshproxywidget import SshProxyWidget
 from exceptions import MountException, NoPubKeyLogin, KnownHost
 from bitbase import URL_ENCRYPT_TRANSITION
 import qttools
-
-
-class SshProxyWidget(QWidget):
-    """Used in SSH snapshot profiles on the General tab.
-
-    Dev note by buhtz (2024-04): Just a quick n dirty solution until the
-    re-design and re-factoring of the whole dialog.
-    """
-    def __init__(self, parent, host, port, user):
-        super().__init__(parent)
-
-        if host == '':
-            port = ''
-            user = ''
-
-        elif isinstance(port, int):
-            port = str(port)
-
-        vlayout = QVBoxLayout(self)
-        # zero margins
-        vlayout.setContentsMargins(0, 0, 0, 0)
-
-        self._checkbox = QCheckBox(_('SSH Proxy'), self)
-        vlayout.addWidget(self._checkbox)
-        self._checkbox.stateChanged.connect(self._slot_checkbox_changed)
-
-        hlayout = QHBoxLayout()
-        vlayout.addLayout(hlayout)
-
-        hlayout.addWidget(QLabel(_('Host:'), self))
-        self.host_edit = QLineEdit(host, self)
-        hlayout.addWidget(self.host_edit)
-
-        hlayout.addWidget(QLabel(_('Port:'), self))
-        self.port_edit = QLineEdit(port, self)
-        hlayout.addWidget(self.port_edit)
-
-        hlayout.addWidget(QLabel(_('User:'), self))
-        self.user_edit = QLineEdit(user, self)
-        hlayout.addWidget(self.user_edit)
-
-        if host == '':
-            self._disable()
-
-        qttools.set_wrapped_tooltip(
-            self,
-            _('Connect to the target host via this proxy (also known as a '
-              'jump host). See "-J" in the "ssh" command documentation or '
-              '"ProxyJump" in "ssh_config" man page for details.')
-        )
-
-    def _slot_checkbox_changed(self, state):
-        if Qt.CheckState(state) == Qt.CheckState.Checked:
-            self._enable()
-        else:
-            self._disable()
-
-    def _set_default(self):
-        """Set GUI elements back to default."""
-        self.host_edit.setText('')
-        self.port_edit.setText('22')
-        self.user_edit.setText(getpass.getuser())
-
-    def _disable(self):
-        self._set_default()
-        self._enable(False)
-
-    def _enable(self, enable=True):
-        # QEdit and QLabel's
-        lay = self.layout().itemAt(1)
-        for idx in range(lay.count()):
-            lay.itemAt(idx).widget().setEnabled(enable)
-
-    def values(self):
-        if self._checkbox.isChecked():
-            return {
-                'host': self.host_edit.text(),
-                'port': self.port_edit.text(),
-                'user': self.user_edit.text(),
-            }
-
-        else:
-            return {
-                'host': '',
-                'port': '',
-                'user': '',
-            }
-
+from editusercallback import EditUserCallback
+from restoreconfigdialog import RestoreConfigDialog
 
 class SettingsDialog(QDialog):
     def __init__(self, parent):
@@ -2170,408 +2085,3 @@ class SettingsDialog(QDialog):
             self.parent.remount(self.originalCurrentProfile,
                                 self.originalCurrentProfile)
             self.parent.updateProfiles()
-
-
-class RestoreConfigDialog(QDialog):
-    """
-    Show a dialog that will help to restore BITs configuration.
-    User can select a config from previous snapshots.
-    """
-
-    def __init__(self, parent):
-        super(RestoreConfigDialog, self).__init__(parent)
-
-        self.parent = parent
-        self.config = parent.config
-        self.snapshots = parent.snapshots
-
-        import icon
-        self.icon = icon
-        self.setWindowIcon(icon.SETTINGS_DIALOG)
-        self.setWindowTitle(_('Import configuration'))
-
-        layout = QVBoxLayout(self)
-        layout.addWidget(self._create_hint_label())
-
-        # treeView
-        self.treeView = qttools.MyTreeView(self)
-        self.treeViewModel = QFileSystemModel(self)
-        self.treeViewModel.setRootPath(QDir().rootPath())
-        self.treeViewModel.setReadOnly(True)
-        self.treeViewModel.setFilter(QDir.Filter.AllDirs |
-                                     QDir.Filter.NoDotAndDotDot |
-                                     QDir.Filter.Hidden)
-
-        self.treeViewFilterProxy = QSortFilterProxyModel(self)
-        self.treeViewFilterProxy.setDynamicSortFilter(True)
-        self.treeViewFilterProxy.setSourceModel(self.treeViewModel)
-
-        self.treeViewFilterProxy.setFilterRegularExpression(r'^[^\.]')
-
-        self.treeView.setModel(self.treeViewFilterProxy)
-        for col in range(self.treeView.header().count()):
-            self.treeView.setColumnHidden(col, col != 0)
-        self.treeView.header().hide()
-
-        # expand users home
-        self.expandAll(os.path.expanduser('~'))
-        layout.addWidget(self.treeView)
-
-        # context menu
-        self.treeView.setContextMenuPolicy(
-            Qt.ContextMenuPolicy.CustomContextMenu)
-        self.treeView.customContextMenuRequested.connect(self.onContextMenu)
-        self.contextMenu = QMenu(self)
-        self.btnShowHidden = self.contextMenu.addAction(
-            icon.SHOW_HIDDEN, _('Show hidden files'))
-        self.btnShowHidden.setCheckable(True)
-        self.btnShowHidden.toggled.connect(self.onBtnShowHidden)
-
-        # colors
-        self.colorRed = QPalette()
-        self.colorRed.setColor(
-            QPalette.ColorRole.WindowText, QColor(205, 0, 0))
-        self.colorGreen = QPalette()
-        self.colorGreen.setColor(
-            QPalette.ColorRole.WindowText, QColor(0, 160, 0))
-
-        # wait indicator which will show that the scan for
-        # snapshots is still running
-        self.wait = QProgressBar(self)
-        self.wait.setMinimum(0)
-        self.wait.setMaximum(0)
-        self.wait.setMaximumHeight(7)
-        layout.addWidget(self.wait)
-
-        # show where a snapshot with config was found
-        self.lblFound = QLabel(_('No config found'), self)
-        self.lblFound.setWordWrap(True)
-        self.lblFound.setPalette(self.colorRed)
-        layout.addWidget(self.lblFound)
-
-        # show profiles inside the config
-        self.widgetProfiles = QWidget(self)
-        self.widgetProfiles.setContentsMargins(0, 0, 0, 0)
-        self.widgetProfiles.hide()
-        self.gridProfiles = QGridLayout()
-        self.gridProfiles.setContentsMargins(0, 0, 0, 0)
-        self.gridProfiles.setHorizontalSpacing(20)
-        self.widgetProfiles.setLayout(self.gridProfiles)
-        layout.addWidget(self.widgetProfiles)
-
-        self.restoreConfig = None
-
-        self.scan = ScanFileSystem(self)
-
-        self.treeView.myCurrentIndexChanged.connect(self.indexChanged)
-        self.scan.foundConfig.connect(self.scanFound)
-        self.scan.finished.connect(self.scanFinished)
-
-        buttonBox = QDialogButtonBox(self)
-        self.restoreButton = buttonBox.addButton(
-            _('Import'), QDialogButtonBox.ButtonRole.AcceptRole)
-        self.restoreButton.setEnabled(False)
-        buttonBox.addButton(QDialogButtonBox.StandardButton.Cancel)
-        buttonBox.accepted.connect(self.accept)
-        buttonBox.rejected.connect(self.reject)
-        layout.addWidget(buttonBox)
-
-        self.scan.start()
-
-        self.resize(600, 700)
-
-    def _create_hint_label(self):
-        """Create the label to explain how and where to find existing config
-        file.
-
-        Returns:
-            (QLabel): The label
-        """
-
-        samplePath = os.path.join(
-            'backintime',
-            self.config.host(),
-            getpass.getuser(), '1',
-            snapshots.SID(datetime.datetime.now(), self.config).sid
-        )
-        samplePath = f'</ br><code>{samplePath}</code>'
-
-        text_a = _(
-            'Select the snapshot folder from which the configuration '
-            'file should be imported. The path may look like: {samplePath}'
-        ).format(samplePath=samplePath)
-
-        text_b = _(
-            'If the folder is located on an external or remote drive, '
-            'it must be manually mounted beforehand.'
-        )
-
-        label = QLabel(f'<p>{text_a}</p><p>{text_b}</p>', self)
-        label.setWordWrap(True)
-
-        return label
-
-    def pathFromIndex(self, index):
-        """
-        return a path string for a given treeView index
-        """
-        sourceIndex = self.treeViewFilterProxy.mapToSource(index)
-        return str(self.treeViewModel.filePath(sourceIndex))
-
-    def indexFromPath(self, path):
-        """
-        return the index for path which can be used in treeView
-        """
-        indexSource = self.treeViewModel.index(path)
-        return self.treeViewFilterProxy.mapFromSource(indexSource)
-
-    def indexChanged(self, current, previous):
-        """
-        called every time a new item is chosen in treeView.
-        If there was a config found inside the selected folder, show
-        available information about the config.
-        """
-        cfg = self.searchConfig(self.pathFromIndex(current))
-        if cfg:
-            self.expandAll(
-                os.path.dirname(os.path.dirname(cfg._LOCAL_CONFIG_PATH)))
-            self.lblFound.setText(cfg._LOCAL_CONFIG_PATH)
-            self.lblFound.setPalette(self.colorGreen)
-            self.showProfile(cfg)
-            self.restoreConfig = cfg
-        else:
-            self.lblFound.setText(_('No config found'))
-            self.lblFound.setPalette(self.colorRed)
-            self.widgetProfiles.hide()
-            self.restoreConfig = None
-        self.restoreButton.setEnabled(bool(cfg))
-
-    def searchConfig(self, path):
-        """
-        try to find config in couple possible subfolders
-        """
-        snapshotPath = os.path.join(
-            'backintime', self.config.host(), getpass.getuser())
-
-        tryPaths = ['', '..', 'last_snapshot']
-        tryPaths.extend([
-            os.path.join(snapshotPath, str(i), 'last_snapshot')
-            for i in range(10)])
-
-        for p in tryPaths:
-            cfgPath = os.path.join(path, p, 'config')
-
-            if os.path.exists(cfgPath):
-
-                try:
-                    cfg = config.Config(cfgPath)
-
-                    if cfg.isConfigured():
-                        return cfg
-
-                except Exception as exc:
-                    logger.error(
-                        f'Unhandled branch in code! See in {__file__} '
-                        f'SettingsDialog.searchConfig()\n{exc}')
-                    pass
-
-        return
-
-    def expandAll(self, path):
-        """
-        expand all folders from filesystem root to given path
-        """
-        paths = [path, ]
-        while len(path) > 1:
-            path = os.path.dirname(path)
-            paths.append(path)
-        paths.append('/')
-        paths.reverse()
-        [self.treeView.expand(self.indexFromPath(p)) for p in paths]
-
-    def showProfile(self, cfg):
-        """
-        show information about the profiles inside cfg
-        """
-        child = self.gridProfiles.takeAt(0)
-
-        while child:
-            child.widget().deleteLater()
-            child = self.gridProfiles.takeAt(0)
-
-        for row, profileId in enumerate(cfg.profiles()):
-
-            for col, txt in enumerate((
-                    _('Profile:') + str(profileId),
-                    cfg.profileName(profileId),
-                    _('Mode:') + cfg.SNAPSHOT_MODES[
-                        cfg.snapshotsMode(profileId)][1]
-                    )):
-                self.gridProfiles.addWidget(QLabel(txt, self), row, col)
-
-        self.gridProfiles.setColumnStretch(col, 1)
-        self.widgetProfiles.show()
-
-    def scanFound(self, path):
-        """
-        scan hit a config. Expand the snapshot folder.
-        """
-        self.expandAll(os.path.dirname(path))
-
-    def scanFinished(self):
-        """
-        scan is done. Delete the wait indicator
-        """
-        self.wait.deleteLater()
-
-    def onContextMenu(self, point):
-        self.contextMenu.exec(self.treeView.mapToGlobal(point))
-
-    def onBtnShowHidden(self, checked):
-        if checked:
-            self.treeViewFilterProxy.setFilterRegularExpression(r'')
-        else:
-            self.treeViewFilterProxy.setFilterRegularExpression(r'^[^\.]')
-
-    def accept(self):
-        """
-        handle over the dict from the selected config. The dict contains
-        all settings from the config.
-        """
-        if self.restoreConfig:
-            self.config.dict = self.restoreConfig.dict
-        super(RestoreConfigDialog, self).accept()
-
-    def exec(self):
-        """
-        stop the scan thread if it is still running after dialog was closed.
-        """
-        ret = super(RestoreConfigDialog, self).exec()
-        self.scan.stop()
-        return ret
-
-
-class ScanFileSystem(QThread):
-    CONFIG = 'config'
-    BACKUP = 'backup'
-    BACKINTIME = 'backintime'
-
-    foundConfig = pyqtSignal(str)
-
-    def __init__(self, parent):
-        super(ScanFileSystem, self).__init__(parent)
-        self.stopper = False
-
-    def stop(self):
-        """
-        prepare stop and wait for finish.
-        """
-        self.stopper = True
-        return self.wait()
-
-    def run(self):
-        """
-        search in order of hopefully fastest way to find the snapshots.
-        1. /home/USER 2. /media 3. /mnt and at last filesystem root.
-        Already searched paths will be excluded.
-        """
-        searchOrder = [os.path.expanduser('~'), '/media', '/mnt', '/']
-        for scan in searchOrder:
-            exclude = searchOrder[:]
-            exclude.remove(scan)
-            for path in self.scanPath(scan, exclude):
-                self.foundConfig.emit(path)
-
-    def scanPath(self, path, excludes=()):
-        """
-        walk through all folders and try to find 'config' file.
-        If found make sure it is nested in backintime/FOO/BAR/1/2345/config and
-        return its path.
-        Exclude all paths from excludes and also
-        all backintime/FOO/BAR/1/2345/backup
-        """
-        for root, dirs, files in os.walk(path, topdown=True):
-
-            if self.stopper:
-                return
-
-            for exclude in excludes:
-                exDir, exBase = os.path.split(exclude)
-
-                if root == exDir:
-
-                    if exBase in dirs:
-                        del dirs[dirs.index(exBase)]
-
-            if self.CONFIG in files:
-                rootdirs = root.split(os.sep)
-
-                if len(rootdirs) > 4 and rootdirs[-5].startswith(self.BACKINTIME):
-
-                    if self.BACKUP in dirs:
-                        del dirs[dirs.index(self.BACKUP)]
-
-                    yield root
-
-
-class EditUserCallback(QDialog):
-    def __init__(self, parent):
-        super(EditUserCallback, self).__init__(parent)
-        self.config = parent.config
-        self.script = self.config.takeSnapshotUserCallback()
-
-        import icon
-        self.setWindowIcon(icon.SETTINGS_DIALOG)
-        self.setWindowTitle(self.script)
-        self.resize(800, 500)
-
-        layout = QVBoxLayout(self)
-        self.edit = QPlainTextEdit(self)
-
-        try:
-            with open(self.script, 'rt') as f:
-                self.edit.setPlainText(f.read())
-
-        except IOError:
-            pass
-
-        layout.addWidget(self.edit)
-
-        btnBox = QDialogButtonBox(
-            QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel,
-            parent=self)
-
-        btnBox.accepted.connect(self.accept)
-        btnBox.rejected.connect(self.reject)
-        layout.addWidget(btnBox)
-
-    def checkScript(self, script):
-        m = re.match(r'^#!(/[\w/-]+)\n', script)
-
-        if not m:
-            logger.error(
-                'user-callback script has no shebang (#!/bin/sh) line.')
-            self.config.errorHandler(
-                'user-callback script has no shebang (#!/bin/sh) line.')
-
-            return False
-
-        if not tools.checkCommand(m.group(1)):
-            logger.error('Shebang in user-callback script is not executable.')
-            self.config.errorHandle(
-                'Shebang in user-callback script is not executable.')
-
-            return False
-
-        return True
-
-    def accept(self):
-        if not self.checkScript(self.edit.toPlainText()):
-            return
-
-        with open(self.script, 'wt') as f:
-            f.write(self.edit.toPlainText())
-
-        os.chmod(self.script, 0o755)
-
-        super(EditUserCallback, self).accept()
